refactoring day 1 names, public and private methods, only exposing the solve method.
