+++ day1/day1_test.go
@@ -1,7 +1,6 @@
-package day1_test
+package day1
 
 import (
-	"aoctdd/day1"
 	"testing"
 )
 
@@ -11,7 +10,7 @@ func TestParseLineFindsTheFirstAndLastNumbersInALineWithTwoNumbers(t *testing.T)
 	t.Parallel()
 	str := "1abc2"
 	want := 12
-	got := day1.ParseLine(str)
+	got := parseLine(str)
 	if want != got {
 		t.Errorf("want %d, got %d", want, got)
 	}
@@ -21,7 +20,7 @@ func TestParseLineFindsTheFirstAndLastNumbersInALineWithTwoNumbersWithCharacters
 	t.Parallel()
 	str := "pqr3stu8vwx"
 	want := 38
-	got := day1.ParseLine(str)
+	got := parseLine(str)
 	if want != got {
 		t.Errorf("want %d, got %d", want, got)
 	}
@@ -31,7 +30,7 @@ func TestParseLineFindsTheFirstAndLastNumbersWithMultipleNumbers(t *testing.T) {
 	t.Parallel()
 	str := "a1b2c3d4e5f"
 	want := 15
-	got := day1.ParseLine(str)
+	got := parseLine(str)
 	if want != got {
 		t.Errorf("want %d, got %d", want, got)
 	}
@@ -41,7 +40,7 @@ func TestParseLineFindsTheFirstAndLastNumbersWithASingleNumber(t *testing.T) {
 	t.Parallel()
 	str := "treb7uchet"
 	want := 77
-	got := day1.ParseLine(str)
+	got := parseLine(str)
 	if want != got {
 		t.Errorf("want %d, got %d", want, got)
 	}
@@ -58,7 +57,7 @@ func TestSumLinesWillAddAllResults(t *testing.T) {
 		"treb7uchet",
 	}
 	for _, line := range lines {
-		got = day1.SumLines(got, line)
+		got = SumLines(got, line)
 	}
 	if want != got {
 		t.Errorf("want %d, got %d", want, got)
@@ -71,7 +70,7 @@ func TestParseDigitStringCanFindValidDigitsWhenThereIsOnlyOne(t *testing.T) {
 	t.Parallel()
 	line := "two"
 	want := 2
-	got, _ := day1.ParseDigitString(line, 0)
+	got, _ := parseDigitString(line, 0)
 	if want != got {
 		t.Errorf("want %d, got %d", want, got)
 	}
@@ -80,7 +79,7 @@ func TestParseDigitStringCanFindValidDigitsWhenThereIsOnlyOne(t *testing.T) {
 func TestParseDigitReturnsAnErrorWhereThereIsNoValidDigitStrings(t *testing.T) {
 	t.Parallel()
 	line := "fail"
-	got, err := day1.ParseDigitString(line, 0)
+	got, err := parseDigitString(line, 0)
 	if err == nil {
 		t.Errorf("expected no valid strings error, got %d", got)
 	}
@@ -90,7 +89,7 @@ func TestParseLineCanFindValidDigitStringsInLine(t *testing.T) {
 	t.Parallel()
 	line := "two1nine"
 	want := 29
-	got := day1.ParseLine(line)
+	got := parseLine(line)
 	if want != got {
 		t.Errorf("want %d, got %d", want, got)
 	}
@@ -110,7 +109,7 @@ func TestParseLineCanFindValidDigitStringsInAllTestLines(t *testing.T) {
 	want := 281
 	got := 0
 	for _, line := range lines {
-		got = day1.SumLines(got, line)
+		got = SumLines(got, line)
 	}
 	if want != got {
 		t.Errorf("want %d, got %d", want, got)
